Improve logic for `grpc_opts` overrides
One major bonus is that `Goth` is only called if an authorization header
isn't supplied, rather than always being invoked to preemptively set a
default value.

Another improvement is that the content-type is forced to
"application/grpc", which is necessary because the Spanner gRPC endpoint
won't work if it's set to anything else.